updated to compile on latest node version

+++ src/node_sigar_format.cc
@@ -100,7 +100,7 @@ SIGAR_DECLARE(char *) sigar_format_size(sigar_uint64_t size, char *buf)
     const char *o = ord;
     int remain;
 
-    if (size == SIGAR_FIELD_NOTIMPL) {
+    if ((int)size == SIGAR_FIELD_NOTIMPL) {
         buf[0] = '-';
         buf[1] = '\0';
         return buf;
